Added char functions
Sorry for the long time before continuing coding this, but i was busy
because of school and things like that.
So, I finally added the access to he car functions and updated the
compiled libraries for MacOSX. The method to get them is the same than
with int and void functions, but the vartype for the functions is funcc
and the function for getting the function is getFunctionc(std::string
nomFunc). Good use of this update.
lacaulac